Created separate folders for client and server side code to improve project organization
