+++ client/src/assets/ceiling.jpg
