+++ client/src/components/CameraControls.jsx
